+++ dart/_remover.pxd
@@ -33,14 +33,14 @@ cdef class _Remover:
 
     # C API
     cdef void _remove(self, Node** node_ptr,
-                      int** X, int* y,
+                      double** X, int* y,
                       int* samples, int n_samples) nogil
 
-    cdef int _check_node(self, Node* node, int** X, int* y,
+    cdef int _check_node(self, Node* node, double** X, int* y,
                           int* samples, int n_samples, int pos_count,
                           SplitRecord *split) nogil
 
-    cdef int _update_splits(self, Node** node_ptr, int** X, int* y,
+    cdef int _update_splits(self, Node** node_ptr, double** X, int* y,
                             int* samples, int n_samples, int pos_count) nogil
 
     cdef void _update_leaf(self, Node** node_ptr, int* y, int* samples,
@@ -49,7 +49,7 @@ cdef class _Remover:
     cdef void _convert_to_leaf(self, Node** node_ptr, int* samples, int n_samples,
                                SplitRecord *split) nogil
 
-    cdef void _retrain(self, Node*** node_ptr, int** X, int* y, int* samples,
+    cdef void _retrain(self, Node*** node_ptr, double** X, int* y, int* samples,
                        int n_samples) nogil
 
     cdef void _get_leaf_samples(self, Node* node, int* remove_samples,
